initial code checkin
The compiler is very basic for this initial commit. Currently, it takes a
filename to compile and outputs llvm assembly to cout. If any errors are
found an error message is given (without line number references) for each
error and no assembly is outputed. The following features are also supported:

= Data Types =
* Several basic data types (void, bool, int8, int16, int32, int64, float, double)
* Auto up-casting type promotion for binary expressions
* Auto exact type matching for assignment
* Most of the basic binary operators (+,-,*,/,%,&,|,^,<<,>>,<,<=,>=,!=,==)
* Support for bool, integer and floating constants

= Functions =
* Support for function declaration and calling
* Argument count and type checking (and promotion)
* Non-void return functions are checked for a return statement
* void function will have a return statment added if missing
* return statements are checked to match the declared type