add support for break, continue, redo
CodeContext has several new functions to store the BasicBlock*
that a LLVM branch statement needs.

	void pushContinueBlock(BasicBlock*)
	void popContinueBlock()
	BasicBlock* getContinueBlock()

	void pushBreakBlock(BasicBlock* block)
	void popBreakBlock()
	BasicBlock* getBreakBlock()

	void pushRedoBlock(BasicBlock* block)
	void popRedoBlock()
	BasicBlock* getRedoBlock()

Loops and other statements needing support for break,continue,redo
need to push a BasicBlock* to one/more of the branch types, and then
pop the blocks when finished. This has been added to the while loop.