add support for ternary operator
The ternary operator will be non-branching if both of the selected
values are non-complex (simple). If either value is complex then a
branch similar to the short-circuit evaluation of the logical and/or
operation will be created.

A simple (non-complex) expression is defined as an expression who's
value needs no more than a memory access to calculate. Currently only
constant values and variable are simple expressions.

Also both types of the selected values must match or an error is logged.