add if statement support
The dangling else problem is solved by making the else keyword left
associative. An else block is always created even if no else is included.
Later optimization passes should remove the unnecessary code. For example,
when the following code is compiled the block labeled 6 is the empty else
block that should be optimized away.

if (x<5)
	x++;
return;

  %1 = load i32* %x
  %2 = icmp slt i32 %1, 5
  br i1 %2, label %3, label %6

; <label>:3			; preds = %0
  %4 = load i32* %x
  %5 = add i32 %4, 1
  store i32 %5, i32* %x
  br label %7

; <label>:6			; preds = %0
  br label %7

; <label>:7			; preds = %6, %3
  ret void