add support for labels and goto statements in functions
This adds support for c-style labels and goto statements. The map
of label names are stored in CodeContext.labelBlocks. The compiler
will check for labels with the same name (in the same function) and
goto statements without a defined label.

There are not checks for goto statements that would create dead code;
nor are there checks for jumping past variable declarations. For example:

	int x = 0;
	goto end;
	int y = 0;
end:
	return y;