add initial basic support for global variables
NGlobalVariableDecl (which extends NVariableDecl) has been added for
declaring global variables. The global variables allow initialization,
but only a single constant value that exactly matches the variable's type.

Global variables in LLVM are declared using GlobalVariable. This required
the SymbolTable and VarTable classes to no longer hold a map of AllocaInst*,
but of Value*. The variables are store by calling CodeContext.storeGlobalVar().