make NStatement::genCode() return void and NExpression::genValue() return Value*
NStatement::genValue() should return void because statements do not return values.
Most classes that were statements were returning nullptr, but this change allows the
compiler to explicitly check that nothing is returned.

NExpression now has genValue() which returns Value* and genCode() has been finalized
with an empty body to prevent sub-classes from accidentally implementing it:

	void genCode(CodeContext& context) final {};

NFunctionPrototype now uses Function* genFunction() instead of genCode(). This makes
it clear that the function's perpose is not for generating code or returning a value,
but to create a function prototype for use by NFunctionDeclaration.