make NodeList extend Node instead of NStatement
Not all NodeLists should be made of NStatements as calling genCode()
on each item in the list might not be enough or correct. For example:

NExpressionList::genCode() calls genValue() which makes it clear that the
value is being descarded. NVariableDeclList::genCode() now takes a type
argument and calls setDataType() on each variable before calling genCode().