fix parser type clash warnings
adding an action to rules where the type changes removes the warning
given by bisonc++. For example, an expression can also be statement,
but using the default action results in a cast warning from bisonc++.

Because NExpressionList is no longer derived from NStatementList (refactoring)
NodeList::copy() has been added to create a copy of a list with
a different node type. The original list should be deleted.