reorder BasicBlocks when calling CodeContext.pushBlock()
The current BasicBlock ordering in a function is dependent on when
the block is created as new blocks are added to the end of the function.
This results in a breath-first block layout which does not mirror
the original source layout. Instead use depth-first layout by moving
blocks to the end of the function when setting them as the current
active block (CodeContext.pushBlock).