wrap llvm::Value* in RValue and LValue
LLVM's Value and Type classes are abstractions of an assembly language.
As such they can only be used as a building block to create the required
high-level language features. For example, there is no unsigned int types.

This adds two wrapper classes around Value* which will be used to
support types not supported by llvm. RValue is for non-named temporary
values; and LValue is for named values.