wrap llvm::Type* in SType*
SType is a wrapper for Type to allow supporting extra high-level types.
The design of SType is similar to Type with a private constructor and
static creator functions. Also just as llvm::Value holds a Type, RValue
now contains a SType.

SType pointers are not unique like Type* are, so can not be compaired
by pointer value -- use matches instead which checks if the Type* is the
same. A type manager should be added into the context to support this.

RValue does not delete the SType in it's destructor, so creating STypes
leaks memory for now. A type manager will be added to fix this.