add TypeManager to manage SType pointers
pointers to SType being created were not being deleted. This adds
TypeManager which now handles creating SType* and storing the
pointer so it doesn't get lost. C++11 shared_ptr is used to make
sure all SType* are freed when ~TypeManager() is called. CodeContext
holds the instance of TypeManager.

TypeManager makes sure that there is only one instance of a SType
created for each unique type. This allows SType* to be compaired
using == like llvm::Type*.