wrap llvm::Function/llvm::FunctionType in SFunction/SFunctionType
This wraps llvm function and function types in our SFunction and
SFunctionType. FunctionManager manages looking up functions by name
and an instance is stored in the CodeContext.

As a bonus this removes many calls to SType::get(CodeContext&, Type*)
which was mainly used because function return types and parameters were
not wrapped in a SType like they are now with SFunctionType. The remaining
calls will be removed later.