add support for vec type
The vec type is a llvm vector type. LLVM usually compiles operations
with vectors to SIMD instruction like SSE. The syntax is:

	vec<size, type>

where size is a constant integer and type is a integer or float type.
Math operations are allowed among vec variables if the type and size
match. There is currently no support for casting vec types.

	vec<4,int> a, b;
	vec<4,float> c, d;

	a += a * b; // OK
	c = b + c; // Invalid