support multi-level loop branch
all NLoopBranch statments now support an optional integer, for example:
	break 3; redo 2; continue 1;

The number determins what nested level of loop statment to operate
on. A value of 1 is the same as no value and 0 is invalid. Nagative
numbers count in reverse order (outer to inner) and are supported by
the code generator but currently not the parser because of a bug.