add unit testing script to test llvm ir output from compiler
unitTest.py is a simple unit testing tool. It takes as input a test
file which is divided into two sections. The top section is the source
code to test and the bottom section is the expected llvm ir. The script
executes the compiler with the test file and diffs the results with
the expected output.

If no arguments are given unitTest.py runs all test, or a list of
strings can be given and the script will search for matching tests.

three commands are currently supported:

-c | --clean
	removes all runtime test files.
-u | --update
	runs the test and overwrites the expected llvm ir
	with the actual. Used to easily update tests.
-d | --dump
	runs the test but does not clean up the temp files.