fix bug in increment expression with unamed variables
context.loadVar(variable->getName()) was used to store the incremented
value back. This will not work for arrays, structs, unions and functions
as the variable loaded with the given name will be incorrect. This
prevented the following examples from working correctly:

	++func1().a;
	func2()[5]++;
	func3()++;

Instead use variable->loadVar(context) which will return a pointer to
the correct memory location (a new alloc in the case of a function call).