remove useless alloca at the start of each function
an i32 alloca was added to every function as the first instruction to
work arround a bug in llvm 3.2 which would cause a segfault if a function
only had a return instruction. It was also assumed that this was required
because clang was adding it to all functions, but new version of clang
no longer do this.