fix RValue's get constant value functions
llvm's constant value functions don't work on all types. For example
ConstantInt::getSigned() expects an int or vec of int. The RValue
functions didn't check for this which usually resulted in the RValue's
SType not matching with the LLVM Type as LLVM would modify the returned
type. Fix this by using the new SType functions getNumberLike() and
getScalar() to make sure the type passed to llvm is acceptable.