fix issue with pointer decrement operator
NIncrement::genValue() always passed a value of 1 to Inst::BinaryOp().
for the normal ++/-- operator this worked because a subtraction op was
created. But for pointers the getelementptr is used and a decrement
requires a -1 value. This is fixed by adding an int value parameter to
RValue::getOne() and passing 1 or -1.