add argument to unitTest.py to rewrite llvm ir files
llvm 3.7 is the first version of a major refactoring to support opaque pointer
types. This causes many of the tests to fail as the IR syntax has changed, but
the expected and actual outputs are still functionally the same. To get arround
this issue the expected and actual llvm ir are reserialized using llvm-as/llvm-dis.

The new command line option that has been added must be the first "files"
argument and has the format:

	+fromVer-toVer

This instructs unitTest.py to convert the expected and actual .ll files to
.bc using the fromVer llvm tools and then convert the .bc file back to .ll
using the toVer llvm tools. For example passing +3.6-3.7 would call:

	llvm-as-3.6 -o testFile.bc testFile.exp
	llvm-dis-3.7 -o testFile.exp testFile.bc
	llvm-as-3.6 -o testFile.bc testFile.ll
	llvm-dis-3.7 -o testFile.ll testFile.bc

This makes sure that both expected and actual are normalized to the same
llvm version so that they can be compaired.