support recursive struct declarations
this allows types with members of their own type

struct Node {
	int data;
	[2]@this lr;
}

this also changes opaque types to be a real llvm opaque type
and not always have an int8 member.

struct/union creation was broken into two phases. First create an
opaque type and second add the body later if it has one.