support array of classes with constructors where the array size is a variable
arrays of known size worked because the size of the array is part of the type,
but with runtime sized arrays the array size is 0. This caused a segmentation
fault because the check for the end of the array always failed.

Fix by updating CGNDataTypeNew::run() to return both the size of the array in
bytes and the runtime size of the array.