support suffix and fullname options for mangling functions
This changes how class functions are mangled. An optional second parameter
to the mangle attribute has been added to select either suffix or fullname
mangling formats with the default being suffix. Using the suffix format
the symbol name becomes ClassName_Suffix and the fullname format uses
the value as is.

For example the following are identical for a class "List":
	#[mangle("List_this2", "true")]
	#[mangle("this2")]